+++ web/dataRepository/productRepository.ts
@@ -1,16 +1,14 @@
 import { Product } from "@/model/product";
 import { ProductMapper } from "@/dataMapper/productMapper";
+import getAllProducts from "@/ServerActions/Products/getAllProducts";
+import getProductById from "@/ServerActions/Products/getProductById";
 
 export class ProductRepository{
     public static getAllProducts(): Promise<Product[]> {
-        return fetch("/api/products")
-            .then(response => response.json())
-            .then(json => json.map((product: any) => ProductMapper.toDomain(product)));
+		return getAllProducts().then((json) => json?.map((product: any) => ProductMapper.toDomain(product)) || []);
     }
 
     public static getProductById(id: number): Promise<Product> {
-        return fetch(`/api/products/${id}`)
-            .then(response => response.json())
-            .then(json => ProductMapper.toDomain(json));
+		return getProductById(id).then((json) => ProductMapper.toDomain(json));
     }
 }
