+++ web/dataRepository/zoneRepository.ts
@@ -1,11 +1,10 @@
 import {Zone} from "@/model/zone";
 import { ZoneMapper } from "@/dataMapper/zoneMapper";
+import { getAllZones } from "@/ServerActions/Zones/getAllZones";
 
 export class ZoneRepository{
     public static getAllZones(): Promise<Zone[]> {
-        return fetch("/api/zones")
-            .then(response => response.json())
-            .then(json => json.map((zone: any) => ZoneMapper.toDomain(zone)));
+		return getAllZones().then((json) => json?.map((zone: any) => ZoneMapper.toDomain(zone)) || []);
     }
 }
