+++ web/package.json
@@ -27,6 +27,7 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "dompurify": "^3.0.9",
+    "fs": "^0.0.1-security",
     "jsdom": "^24.0.0",
     "next": "14.1.0",
     "next-themes": "^0.2.1",
